Turns tileMap into a class
